requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
numpy==1.26.3
pandas==2.1.4
jinja2==3.1.3
aiofiles==23.2.1
tqdm==4.66.1
